+++ changelog.md
@@ -1,3 +1,14 @@
+### v0.9.0-beta2
+
+* 2018/5/22
+* API "Encode6Bit" is renamed to "EncodeForPrint"
+* API "Decode6Bit" is renamed to "DecodeForPrint"
+
+### v0.9.0-beta1
+
+* 2018/5/22
+* No change
+
 ### v0.9.0-alpha2
 
 * 2018/5/21